Add work data to images mapping
