+++ common/elasticsearch/src/main/scala/uk/ac/wellcome/elasticsearch/ElasticsearchIndexCreator.scala
@@ -43,6 +43,10 @@ class ElasticsearchIndexCreator(
           // If/when we index more records, we should revisit this.
           //
           .shards(1)
+          // Elasticsearch has a default maximum number of fields of 1000.
+          // Because images have all of the WorkData fields defined twice in the mapping,
+          // they end up having more than 1000 fields, so we increase them to 2000
+          .settings(Map("mapping.total_fields.limit"-> 2000))
       }
       .flatMap { response: Response[CreateIndexResponse] =>
         if (response.isError) {
