+++ common/elasticsearch/src/main/scala/uk/ac/wellcome/elasticsearch/ImagesIndexConfig.scala
@@ -1,30 +1,27 @@
 package uk.ac.wellcome.elasticsearch
 
-import com.sksamuel.elastic4s.ElasticDsl._
+import com.sksamuel.elastic4s.ElasticDsl.{keywordField, _}
 import com.sksamuel.elastic4s.requests.analysis.Analysis
-import com.sksamuel.elastic4s.requests.mappings.MappingDefinition
+import com.sksamuel.elastic4s.requests.mappings.{MappingDefinition, ObjectField}
 import com.sksamuel.elastic4s.requests.mappings.dynamictemplate.DynamicMapping
 
 case object ImagesIndexConfig extends IndexConfig {
 
   override val analysis: Analysis = Analysis(List())
 
-  val fullText = textField("fullText")
-  def englishTextField(name: String) =
-    textField(name).fields(
-      keywordField("keyword"),
-      textField("english").analyzer("english")
-    )
-
   val inferredData = objectField("inferredData").fields(
     denseVectorField("features1", 2048),
     denseVectorField("features2", 2048),
     keywordField("lshEncodedFeatures"))
 
-  val source = objectField("source").fields(
+  private def sourceWork(canonicalWork: String): ObjectField = objectField(canonicalWork).fields(
     id("id"),
-    keywordField("type"),
-    keywordField("ontologyType")
+    WorksIndexConfig.data(textField("path")),
+    keywordField("ontologyType"),
+    keywordField("type")
+  )
+  val source = objectField("source").fields(
+    sourceWork("canonicalWork"), sourceWork("redirectedWork"), keywordField("type")
   )
 
   override val mapping: MappingDefinition = properties(
@@ -32,7 +29,6 @@ case object ImagesIndexConfig extends IndexConfig {
     version,
     location("location"),
     source,
-    fullText,
     inferredData
   ).dynamic(DynamicMapping.Strict)
 }
