+++ common/elasticsearch/src/main/scala/uk/ac/wellcome/elasticsearch/IndexConfig.scala
@@ -1,28 +1,8 @@
 package uk.ac.wellcome.elasticsearch
 
-import com.sksamuel.elastic4s.ElasticDsl.{
-  booleanField,
-  dateField,
-  intField,
-  keywordField,
-  objectField,
-  properties,
-  textField,
-  tokenCountField
-}
-import com.sksamuel.elastic4s.requests.analysis.{
-  Analysis,
-  CustomAnalyzer,
-  PathHierarchyTokenizer
-}
-import com.sksamuel.elastic4s.requests.mappings.dynamictemplate.DynamicMapping
-import com.sksamuel.elastic4s.requests.mappings.{
-  FieldDefinition,
-  MappingDefinition,
-  ObjectField
-}
-import uk.ac.wellcome.elasticsearch.ImagesIndexConfig.englishTextField
-
+import com.sksamuel.elastic4s.ElasticDsl.{intField, keywordField, objectField, textField}
+import com.sksamuel.elastic4s.requests.analysis.Analysis
+import com.sksamuel.elastic4s.requests.mappings.MappingDefinition
 trait IndexConfig {
   val mapping: MappingDefinition
   val analysis: Analysis
@@ -37,6 +17,12 @@ trait IndexConfig {
 
   def keywordWithText(name: String) =
     keywordField(name).fields(textField("text"))
+
+  def englishTextField(name: String) =
+    textField(name).fields(
+      keywordField("keyword"),
+      textField("english").analyzer("english")
+    )
 
   def sourceIdentifierFields = Seq(
     keywordField("ontologyType"),
@@ -92,194 +78,3 @@ trait IndexConfig {
   val version = intField("version")
 }
 
-case object WorksIndexConfig extends IndexConfig {
-  // Analysis
-  val pathTokenizer = PathHierarchyTokenizer("path_hierarchy_tokenizer")
-  val pathAnalyzer =
-    CustomAnalyzer("path_hierarchy_analyzer", pathTokenizer.name, Nil, Nil)
-  val analysis = Analysis(
-    analyzers = List(
-      pathAnalyzer
-    ),
-    tokenizers = List(pathTokenizer))
-
-  val sourceIdentifier = objectField("sourceIdentifier")
-    .fields(sourceIdentifierFields)
-
-  val otherIdentifiers = objectField("otherIdentifiers")
-    .fields(sourceIdentifierFields)
-
-  val workType = objectField("workType")
-    .fields(
-      label,
-      keywordField("ontologyType"),
-      keywordField("id")
-    )
-
-  val notes = objectField("notes")
-    .fields(
-      keywordField("type"),
-      englishTextField("content")
-    )
-
-  val period = Seq(
-    label,
-    id(),
-    keywordField("ontologyType"),
-    objectField("range").fields(
-      label,
-      dateField("from"),
-      dateField("to"),
-      booleanField("inferred")
-    )
-  )
-
-  val place = Seq(
-    label,
-    id()
-  )
-
-  val concept = Seq(
-    label,
-    id(),
-    keywordField("ontologyType"),
-    keywordField("type")
-  )
-
-  val agent = Seq(
-    label,
-    id(),
-    keywordField("type"),
-    keywordField("prefix"),
-    keywordField("numeration"),
-    keywordField("ontologyType")
-  )
-
-  val rootConcept = concept ++ agent ++ period
-
-  val subject: Seq[FieldDefinition] = Seq(
-    id(),
-    label,
-    keywordField("ontologyType"),
-    objectField("concepts").fields(rootConcept)
-  )
-
-  def subjects: ObjectField = objectField("subjects").fields(subject)
-
-  def genre(fieldName: String) = objectField(fieldName).fields(
-    label,
-    keywordField("ontologyType"),
-    objectField("concepts").fields(rootConcept)
-  )
-
-  def labelledTextField(fieldName: String) = objectField(fieldName).fields(
-    label,
-    keywordField("ontologyType")
-  )
-
-  def period(fieldName: String) = labelledTextField(fieldName)
-
-  def items(fieldName: String) = objectField(fieldName).fields(
-    id(),
-    location(),
-    englishTextField("title"),
-    keywordField("ontologyType")
-  )
-
-  def englishTextField(name: String) =
-    textField(name).fields(
-      textField("english").analyzer("english")
-    )
-
-  val language = objectField("language").fields(
-    label,
-    keywordField("id"),
-    keywordField("ontologyType")
-  )
-
-  val contributors = objectField("contributors").fields(
-    id(),
-    objectField("agent").fields(agent),
-    objectField("roles").fields(
-      label,
-      keywordField("ontologyType")
-    ),
-    keywordField("ontologyType")
-  )
-
-  val production: ObjectField = objectField("production").fields(
-    label,
-    objectField("places").fields(place),
-    objectField("agents").fields(agent),
-    objectField("dates").fields(period),
-    objectField("function").fields(concept),
-    keywordField("ontologyType")
-  )
-
-  val mergeCandidates = objectField("mergeCandidates").fields(
-    objectField("identifier").fields(sourceIdentifierFields),
-    keywordField("reason")
-  )
-
-  val images = objectField("images").fields(
-    id(),
-    location("location"),
-    version
-  )
-
-  val data: ObjectField =
-    objectField("data").fields(
-      otherIdentifiers,
-      mergeCandidates,
-      workType,
-      textField("title").fields(
-        keywordField("keyword"),
-        textField("english").analyzer("english")
-      ),
-      englishTextField("alternativeTitles"),
-      englishTextField("description"),
-      englishTextField("physicalDescription"),
-      englishTextField("lettering"),
-      objectField("createdDate").fields(period),
-      contributors,
-      subjects,
-      genre("genres"),
-      items("items"),
-      production,
-      language,
-      location("thumbnail"),
-      textField("edition"),
-      notes,
-      intField("duration"),
-      booleanField("merged"),
-      objectField("collectionPath").fields(
-        label,
-        objectField("level").fields(keywordField("type")),
-        textField("path")
-          .copyTo("data.collectionPath.depth")
-          .analyzer(pathAnalyzer.name)
-          .fields(keywordField("keyword")),
-        tokenCountField("depth").analyzer("standard")
-      ),
-      images
-    )
-
-  val fields: Seq[FieldDefinition with Product with Serializable] =
-    Seq(
-      canonicalId,
-      keywordField("ontologyType"),
-      version,
-      sourceIdentifier,
-      objectField("redirect")
-        .fields(sourceIdentifier, canonicalId),
-      keywordField("type"),
-      data,
-      objectField("invisibilityReasons").fields(
-        keywordField("type"),
-        keywordField("info")
-      )
-    )
-
-  val mapping = properties(fields).dynamic(DynamicMapping.Strict)
-
-}
